Revert 'use-debounce' hook and get rid of the autosave for now
